requirements.txt
flask~=2.3.3
psutil~=5.9.5
python-nmap~=0.7.1